DOP-5598: Updated pathing of toc.ts (#1414)
