+++ src/components/Sidenav/DocsHomeButton.js
@@ -10,6 +10,7 @@ import useViewport from '../../hooks/useViewport';
 import { baseUrl } from '../../utils/base-url';
 import { theme } from '../../theme/docsTheme';
 import useScreenSize from '../../hooks/useScreenSize';
+import { getFeatureFlags } from '../../utils/feature-flags';
 import { sideNavItemBasePadding } from './styles/sideNavItem';
 import { titleStyle, logoLinkStyling } from './styles/sideNavItem';
 
@@ -31,9 +32,10 @@ const homeLinkStyle = LeafyCSS`
   }
 `;
 
-const containerStyle = LeafyCSS`
+const containerStyle = (isUnifiedToc, isDesktop) => LeafyCSS`
   display: flex;
   align-items: center;
+  ${isUnifiedToc && !isDesktop && 'width : 162px'}
 `;
 
 const logoTextStyling = LeafyCSS`
@@ -47,8 +49,9 @@ const logoTextStyling = LeafyCSS`
 
 const DocsHomeButton = () => {
   const viewport = useViewport(false);
-  const { isTabletOrMobile } = useScreenSize();
+  const { isTabletOrMobile, isDesktop } = useScreenSize();
   const { darkMode } = useDarkMode();
+  const { isUnifiedToc } = getFeatureFlags();
 
   const sideNavHome = useMemo(
     () => (
@@ -75,7 +78,7 @@ const DocsHomeButton = () => {
     [darkMode]
   );
   return (
-    <div className={cx(containerStyle)}>
+    <div className={cx(containerStyle(isUnifiedToc, isDesktop))}>
       {!isTabletOrMobile && viewport.scrollY > parseInt(theme.header.navbarHeight, 10) ? homeNav : sideNavHome}
     </div>
   );
