+++ src/utils/assert-trailing-and-leading-slash.ts
@@ -1,0 +1,8 @@
+import { assertLeadingSlash } from './assert-leading-slash';
+import { assertTrailingSlash } from './assert-trailing-slash';
+
+export const assertLeadingAndTrailingSlash = (url: string) => {
+  return assertLeadingSlash(assertTrailingSlash(url));
+};
+
+// export default assertLeadingAndTrailingSlash;
